Return statement

+++ Bachelor-Thesis/src/Core/Chunk.h
@@ -16,16 +16,16 @@ namespace Lang {
 	};
 
 	struct Chunk {
-		std::vector<uint8_t> Code;
+		std::vector<uint16_t> Code;
 		std::vector<uint16_t> Lines;
 		std::vector<Value> Constants;
 
 		/// <summary>Writes a byte to the bytcode.</summary>
-		void Write(uint8_t byte, uint16_t line);
+		void Write(uint16_t byte, uint16_t line);
 
 		/// <summary>Adds a new value to the constants array.</summary>
 		/// <returns>The index at which the value was added.</returns>
-		uint8_t AddConstant(Value value);
+		uint16_t AddConstant(Value value);
 	};
 
 }
