+++ Bachelor-Thesis/src/Core/Value.h
@@ -5,12 +5,12 @@
 namespace Lang {
 
 	struct Value {
-		uint8_t Dim;
-		std::vector<uint8_t> Shape;
+		uint16_t Dim;
+		std::vector<uint16_t> Shape;
 		std::vector<double> Values;
 
 		Value(double value);
-		Value(uint8_t dim, std::vector<uint8_t> shape, std::vector<double> values);
+		Value(uint16_t dim, std::vector<uint16_t> shape, std::vector<double> values);
 
 		Value operator!();
 		Value operator!=(Value other);
